+++ scriptabit/utility_functions.py
@@ -148,10 +148,14 @@ class UtilityFunctions(object):
         print()
         logging.getLogger(__name__).debug('Running test function')
         print("--------------------")
+        task = self.__hs.create_task(
+            {'text':'new task'},
+            task_type=HabiticaTaskTypes.habits)
+        pprint(task)
         # pprint(self.__hs.get_user())
         # tasks = self.__hs.get_tasks()
-        tasks = self.__hs.get_tasks(task_type=HabiticaTaskTypes.dailies)
-        for t in tasks:
-            pprint(t['type'])
+        # tasks = self.__hs.get_tasks(task_type=HabiticaTaskTypes.dailies)
+        # for t in tasks:
+            # pprint(t['type'])
         print("--------------------")
         print()